+++ src/reactions/component/not-logged-in-popup.jsx
@@ -4,7 +4,7 @@ import OpenStadComponent from '../../component/index.jsx';
 import OpenStadComponentLibs from '../../libs/index.jsx';
 import OpenStadComponentModalPopup from '../../modal-popup/index.jsx';
 
-export default class OpenStadComponentReactionForm extends OpenStadComponent {
+export default class OpenStadComponentNotLoggedInPopup extends OpenStadComponent {
 
   constructor(props) {
     super(props, {
@@ -15,10 +15,12 @@ export default class OpenStadComponentReactionForm extends OpenStadComponent {
   showPopup() {
 		// dispatch an event
     let self = this;
+    let loginUrl = OpenStadComponentLibs.auth.getLoginUrl(self.config);
 		var event = new window.CustomEvent('osc-show-modal-popup', { detail: {
+      divId: self.divId,
       title: 'Login om verder te gaan',
       text: 'Login om een argument te plaatsen, te reageren of een argument te liken',
-      buttonAction: () => { OpenStadComponentLibs.localStorage.set('osc-login-pending-scroll-to-reactions', true); OpenStadComponentLibs.localStorage.set('osc-login-pending-show-details', self.config.ideaId); document.location.href = this.config.loginUrl; },
+      buttonAction: () => { OpenStadComponentLibs.localStorage.set('osc-login-pending-scroll-to-reactions', true); OpenStadComponentLibs.localStorage.set('osc-login-pending-show-details', self.config.ideaId); document.location.href = loginUrl; },
       buttonText: 'Inloggen',
     }});
 		document.dispatchEvent(event);
@@ -26,7 +28,7 @@ export default class OpenStadComponentReactionForm extends OpenStadComponent {
   
   render() {
     return (
-      <OpenStadComponentModalPopup ref={el => (self.modalPopup = el)}/>
+      <OpenStadComponentModalPopup id={this.divId} ref={el => (self.modalPopup = el)}/>
     );
   }