+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.33.0
+* Fix error when clicking on new idea button in mobile view
+* Make votebuttons in idea-details configurable
+* Fix login redirects
+
+## 0.32.0
+* Add new configuration options to reactions
+
 ## 0.31.0
 * Fix closed reactions for moderators
 * Empty value in radio buttons means 'no effect'